Move CSV unit test to importUtils
